+++ src/Controller/ItemDeleteController.php
@@ -3,9 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Item;
-use App\Entity\Project;
-use App\Form\ItemType;
-use App\Repository\ItemRepository;
+use App\Security\Voter\ProjectVoter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -18,9 +16,14 @@ class ItemDeleteController extends AbstractController
 {
     /**
      * @Route("/{item}/delete", name="item_delete", methods={"DELETE"})
+     * @param Request $request
+     * @param Item $item
+     * @return Response
      */
-    public function delete(Request $request, Project $project, Item $item): Response
+    public function delete(Request $request, Item $item): Response
     {
+        $this->denyAccessUnlessGranted(ProjectVoter::PROJECT_EDIT, $item->getProject());
+        $project = $item->getProject();
         if ($this->isCsrfTokenValid('delete'.$item->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($item);
