+++ src/Controller/SceneDeleteController.php
@@ -2,10 +2,8 @@
 
 namespace App\Controller;
 
-use App\Entity\Chapter;
 use App\Entity\Scene;
-use App\Form\SceneType;
-use App\Repository\SceneRepository;
+use App\Security\Voter\ProjectVoter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -18,9 +16,14 @@ class SceneDeleteController extends AbstractController
 {
     /**
      * @Route("/scene/{scene}/delete", name="scene_delete", methods={"DELETE"})
+     * @param Request $request
+     * @param Scene $scene
+     * @return Response
      */
-    public function delete(Request $request, Chapter $chapter, Scene $scene): Response
+    public function delete(Request $request, Scene $scene): Response
     {
+        $this->denyAccessUnlessGranted(ProjectVoter::PROJECT_EDIT, $scene->getChapter()->getProject());
+        $chapter = $scene->getChapter();
         if ($this->isCsrfTokenValid('delete'.$scene->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($scene);
